Relocate task credential creation into separate module
